(REF) Extract BasicArrayAccessTrait from LocoEvent. Split apart PHP 7.x and 8.x variants.
